Extract image-related util functions into module
